backend/requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
xlsxwriter
python-multipart
gspread
google-auth
PyJWT
apify-client
requests
python-calamine

# Database Dependencies
sqlalchemy
psycopg2-binary
python-dotenv
